backend/requirements.txt
google-genai
fastapi[standard]
uvicorn
python-dotenv